feat(image/engine): Move opencv engine to its own file.
